refactor(app): rename and split code
