+++ app-near/src/app/_components/survey/steps/step2/broadcasting/BroadcastingButton.spec.tsx
@@ -3,7 +3,7 @@ import BroadcastingButton from "./BroadcastingButton";
 import {
   type SurveyType,
   type BroadcastType,
-} from "../../types/enums/broadcasting";
+} from "../../../../../../types/enums/broadcasting";
 
 jest.mock("../../env", () => ({
   env: {
