+++ app-near/src/app/_components/survey/steps/step2/broadcasting/BroadcastingPage.tsx
@@ -1,4 +1,4 @@
-import { type SurveyType } from "../../types/enums/broadcasting";
+import { type SurveyType } from "../../../../../../types/enums/broadcasting";
 import BroadcastingButton from "./BroadcastingButton";
 
 interface BroadcastingPageProps {
